src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --all-extras pyproject.toml -o src/requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via fastapi-template (pyproject.toml)
fastapi==0.128.4
    # via fastapi-template (pyproject.toml)
fastapi-cli==0.0.20
    # via fastapi-template (pyproject.toml)
greenlet==3.3.1
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via fastapi-template (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.3.0
    # via pytest
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
packaging==26.0
    # via pytest
pluggy==1.6.0
    # via pytest
psycopg2==2.9.11
    # via fastapi-template (pyproject.toml)
pyasn1==0.6.2
    # via
    #   python-jose
    #   rsa
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pytest==9.0.2
    # via fastapi-template (pyproject.toml)
python-dotenv==1.2.1
    # via
    #   fastapi-template (pyproject.toml)
    #   uvicorn
python-jose==3.5.0
    # via fastapi-template (pyproject.toml)
python-multipart==0.0.22
    # via fastapi-template (pyproject.toml)
pyyaml==6.0.3
    # via uvicorn
rich==14.3.2
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.18.1
    # via fastapi-cli
rsa==4.9.1
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.17.0
    # via ecdsa
sqlalchemy==2.0.46
    # via fastapi-template (pyproject.toml)
starlette==0.52.1
    # via fastapi
typer==0.21.1
    # via fastapi-cli
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.40.0
    # via
    #   fastapi-template (pyproject.toml)
    #   fastapi-cli
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
